Use common functions for rates, info, referral, and notification servers
