+++ src/actions/DeviceReferralActions.js
@@ -2,12 +2,11 @@
 
 import { asArray, asObject, asOptional, asString } from 'cleaners'
 
-import { config } from '../theme/appConfig.js'
 import { type Dispatch, type GetState } from '../types/reduxTypes.js'
 import { type DeviceReferral } from '../types/ReferralTypes.js'
 import { asCurrencyCode, asMessageTweak, asPluginTweak } from '../types/TweakTypes.js'
+import { fetchReferral } from '../util/network'
 import { logEvent } from '../util/tracking.js'
-import { fetchWaterfall } from '../util/utils.js'
 
 const DEVICE_REFERRAL_FILE = 'utilityServer.json'
 
@@ -33,7 +32,7 @@ export const loadDeviceReferral = () => async (dispatch: Dispatch, getState: Get
   // Now try the network:
   try {
     console.log('Fetching app install reason')
-    const reply = await fetchWaterfall(config.referralServers, 'ref')
+    const reply = await fetchReferral('ref')
     if (!reply.ok) {
       throw new Error(`Util server returned status code ${reply.status}`)
     }
