+++ src/components/scenes/CryptoExchangeScene.js
@@ -18,7 +18,8 @@ import { connect } from '../../types/reactRedux.js'
 import { type GuiCurrencyInfo, emptyCurrencyInfo } from '../../types/types.js'
 import { getTokenId } from '../../util/CurrencyInfoHelpers'
 import { getWalletFiat, getWalletName } from '../../util/CurrencyWalletHelpers.js'
-import { DECIMAL_PRECISION, fetchInfo, getDenomFromIsoCode, zeroString } from '../../util/utils'
+import { fetchInfo } from '../../util/network'
+import { DECIMAL_PRECISION, getDenomFromIsoCode, zeroString } from '../../util/utils'
 import { SceneWrapper } from '../common/SceneWrapper.js'
 import { type WalletListResult, WalletListModal } from '../modals/WalletListModal.js'
 import { Airship, showError } from '../services/AirshipInstance'
